admin system added, login system added, reservations system now linked to json file, added view, delete and add reservation options

+++ Reservations.cs
@@ -1,10 +1,13 @@
 ﻿using System;
-using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using Newtonsoft.Json.Linq;
+using projectb;
 
 namespace testproject1 {
     public class Reservations {
 
-        static IList<string> reservations = new List<string>();
+        private static string reservationsDatabase = Directory.GetCurrentDirectory() + "/../../../json_files/reservations.json";
 
         public static void ReservationSystem() {
 
@@ -18,7 +21,9 @@ namespace testproject1 {
 
             Random random = new Random();
 
-            //FUNCTIONS
+            //---------FUNCTIONS-------------
+
+            //Get name for reservation.
             string getName() {
                 Console.WriteLine("Please enter a name for the reservation ");
                 reservationName = Console.ReadLine();
@@ -29,6 +34,7 @@ namespace testproject1 {
                 return reservationName;
             }
 
+            //Get time for reservation.
             string getTime() {
                 Console.WriteLine("Please enter the time of the reservation (ex: 18:00)");
                 reservationTime = Console.ReadLine();
@@ -39,6 +45,7 @@ namespace testproject1 {
                 return reservationTime;
             }
 
+            //Get amount of guests for reservation.
             string getAmount() {
                 Console.WriteLine("For how many people do you want to reserve? ");
                 reservationAmount = Console.ReadLine();
@@ -49,52 +56,115 @@ namespace testproject1 {
                 return reservationAmount;
             }
 
+            //Generate a reservation code to link to the reservation.
             int generateCode() {
                 return random.Next(1, 1000);
             }
 
+            //Reads inputs for the reservation menu's options.
             void readMenuInput() {
                 menuSelection = Console.ReadLine();
             }
 
+            //Makes a reservation and saves it to the json file reservations.json
             void makeReservation() {
-                string name = getName();
-                string time = getTime();
-                string amount = getAmount();
-                int code = generateCode();
-                string reservation = "Code: #" + code + "    Name: " + name + "  Time: " + time + "  Amount of Guests: " + amount;
-                reservations.Add(reservation);
-                Console.WriteLine("...Reservation Saved...");
+                var name = getName();
+                var time = getTime();
+                var amount = getAmount();
+                var code = generateCode();
+
+                var newReservation = "{ 'name': '" + name + "', 'time': '" + time + "', 'amount': " + amount + ", 'code': " + code + "}";
+                try {
+                    var json = File.ReadAllText(reservationsDatabase);
+                    var jsonObj = JObject.Parse(json);
+                    var reservationsArray = jsonObj.GetValue("reservations") as JArray;
+                    var makeReservation = JObject.Parse(newReservation);
+                    reservationsArray.Add(makeReservation);
+
+                    jsonObj["reservations"] = reservationsArray;
+                    string newJsonResult = Newtonsoft.Json.JsonConvert.SerializeObject(jsonObj,
+                                           Newtonsoft.Json.Formatting.Indented);
+                    File.WriteAllText(reservationsDatabase, newJsonResult);
+                }
+                catch (Exception ex) {
+                    Console.WriteLine("Add Error : " + ex.Message.ToString());
+                }
+
+                Console.WriteLine("Reservation Saved... Your reservation code is: " + code);
                 Console.WriteLine("What would you like to do now? (enter \'help\' to see options)");
             }
 
+            //View all of the reservations in the reservations.json file
             void viewReservation() {
                 Console.WriteLine("---------- ALL RESERVATIONS ----------");
-                foreach (var el in reservations)
-                    Console.WriteLine(el);
+                var json = File.ReadAllText(reservationsDatabase);
+                try {
+                    var jObject = JObject.Parse(json);
+
+                    if (jObject != null) {
+                        JArray reservationsArray = (JArray)jObject["reservations"];
+                        if (reservationsArray != null) {
+                            foreach (var item in reservationsArray) {
+                                Console.WriteLine("Code: " + item["code"].ToString() + "    Name: " + item["name"].ToString() + "   Time: " + item["time"].ToString() + "   Amount: " + item["amount"].ToString());
+                            }
+
+                        }
+                    }
+                }
+                catch (Exception) {
+
+                    throw;
+                }
+               
                 Console.WriteLine("\nWhat would you like to do now? (enter \'help\' to see options)");
             }
 
+            //Delete a reservation from the reservations.json file by entering the reservation code.
             void deleteReservation() {
-                viewReservation();
-                Console.WriteLine("Enter the code of the reservation to delete it: ");
-                string toDelete = Console.ReadLine();
-                while (string.IsNullOrEmpty(toDelete)) {
-                    Console.WriteLine("Empty input, Please try again");
-                    toDelete = Console.ReadLine();
-                }
-                for (int el = reservations.Count - 1; el >= 0; el--) {
-                    if (reservations[el].ToString().Contains(toDelete)) {
-                        reservations.RemoveAt(el);
-                        Console.WriteLine("...Reservation Succesfully Deleted...");
+                var json = File.ReadAllText(reservationsDatabase);
+                try {
+                    var jObject = JObject.Parse(json);
+                    JArray reservationsArray = (JArray)jObject["reservations"];
+                    Console.Write("\nEnter Reservation Code to Delete Reservation: ");
+                    var readResCode = Console.ReadLine();
+
+                    if (int.TryParse(readResCode, out _)) {
+                        var resCode = Convert.ToInt32(readResCode);
+                        if (resCode > 0 && json.ToString().Contains(resCode.ToString())) {
+                            var name = string.Empty;
+                            var resToDeleted = reservationsArray.FirstOrDefault(obj => obj["code"].Value<int>() == resCode);
+
+                            reservationsArray.Remove(resToDeleted);
+
+                            string output = Newtonsoft.Json.JsonConvert.SerializeObject(jObject, Newtonsoft.Json.Formatting.Indented);
+                            File.WriteAllText(reservationsDatabase, output);
+                            Console.WriteLine("Reservation has been deleted.");
+                        }
+                        else {
+                            Console.WriteLine("Reservation not found, please try again");
+                            deleteReservation();
+                        }
+                    } else {
+                        Console.Write("Invalid Reservation Code, Try Again!\n");
+                        deleteReservation();
                     }
-                    Console.WriteLine("\nWhat would you like to do now? (enter \'help\' to see options)");
+                }
+                catch (Exception) {
+
+                    throw;
                 }
             }
 
+            //Prints out the options for the reservation menu. If admin is logged in print full menu options.
+            //else print guest menu options.
             void menuHelp() {
                 Console.WriteLine("\n----Reservation Menu----");
-                Console.WriteLine(" [M] - Make Reservation\n [V] - View Reservations\n [D] - Delete Reservations\n [E] - Exit and back to the main page\n");
+                if (AdminSystem.adminLoggedin == true) {
+                    Console.WriteLine(" [M] - Make Reservation\n [V] - View Reservations\n [D] - Delete Reservations\n [E] - Exit and back to the main page\n");
+                }
+                else {
+                    Console.WriteLine(" [M] - Make Reservation\n [E] - Exit and back to the main page\n");
+                }
             }
 
             //Reservation Menu
@@ -103,34 +173,57 @@ namespace testproject1 {
 
             while (menuRunning) {
                 readMenuInput();
-                switch (menuSelection.ToLower()) {
-                    case "m":
-                        makeReservation();
-                        break;
-                    case "v":
-                        viewReservation();
-                        break;
-                    case "d":
-                        deleteReservation();
-                        break;
-                    case "help":
-                        menuHelp();
-                        break;
-                    case "e":
-                        Console.WriteLine("\n----Welcome Back To The Main Menu----");
-                        Console.WriteLine("Enter \'help\' to view the options!");
-                        menuRunning = false;
-                        break;
-                    case null:
-                    case "":
-                        Console.WriteLine("Empty input, please try again");
-                        break;
-                    default:
-                        Console.WriteLine("Not a valid input, please try again.");
-                        break;
+                if (AdminSystem.adminLoggedin == true) {
+                    switch (menuSelection.ToLower()) {
+                        case "m":
+                            makeReservation();
+                            break;
+                        case "v":
+                            viewReservation();
+                            break;
+                        case "d":
+                            deleteReservation();
+                            break;
+                        case "help":
+                            menuHelp();
+                            break;
+                        case "e":
+                            Console.WriteLine("\n----Welcome Back To The Main Menu----");
+                            Console.WriteLine("Enter \'help\' to view the options!");
+                            menuRunning = false;
+                            break;
+                        case null:
+                        case "":
+                            Console.WriteLine("Empty input, please try again");
+                            break;
+                        default:
+                            Console.WriteLine("Not a valid input, please try again.");
+                            break;
+                    }
                 }
-            }
-            
+                else { 
+                    switch (menuSelection.ToLower()) {
+                        case "m":
+                            makeReservation();
+                            break;
+                        case "help":
+                            menuHelp();
+                            break;
+                        case "e":
+                            Console.WriteLine("\n----Welcome Back To The Main Menu----");
+                            Console.WriteLine("Enter \'help\' to view the options!");
+                            menuRunning = false;
+                            break;
+                        case null:
+                        case "":
+                            Console.WriteLine("Empty input, please try again");
+                            break;
+                        default:
+                            Console.WriteLine("Not a valid input, please try again.");
+                            break;
+                    }
+                }
+            } 
         }
     }
 }
